+++ .eslintrc.js
@@ -88,6 +88,7 @@ module.exports = {
         peer: 'writable',
         settings: 'writable',
         meet: 'writable',
+        meetLowLevel: 'writable',
         meetHighLevel: 'writable',
         zoneManager: 'writable',
         notificationMessage: 'writable',
