+++ core/modules/meet/client/meet-low-level.hbs.html
@@ -1,47 +1,3 @@
-<template name="remoteVideoTrack">
-    <div class="remote-stream">
-        <div class="stream-name">{{displayName}}</div>
-        <div class="webcam">
-            {{#if isMuted}}
-            <img src="{{avatar}}" alt="avatar" class="avatar" />
-            {{/if}}
-            <video id="{{track.getParticipantId}}-camera" type="video/mp4" autoplay></video>
-            {{#if desktopTrack}} {{> remoteDesktopTrack track=desktopTrack}} {{/if}}
-        </div>
-    </div>
-</template>
-
-<template name="remoteDesktopTrack">
-    <div class="js-screenshare screenshare-preview {{#if isMuted}}hide{{/if}}">
-        <video id="{{track.getParticipantId}}-desktop" type="video/mp4" autoplay muted></video>
-    </div>
-</template>
-
-<template name="remoteAudioTrack">
-    <div class="remote-audio">
-        {{#if isMuted}}
-        <div class="mute">
-            <div>🎤</div>
-            <div>❌</div>
-        </div>
-        {{/if}}
-        <audio autoplay="1" id="{{track.getParticipantId}}-audio"></audio>
-    </div>
-</template>
-
-<template name="remoteFullscreen">
-    <div class="fullscreen-container">
-        <div class="fullscreen">
-            <div class="fullscreen-background">
-                <video id="{{track.getParticipantId}}-fullscreenBackground" type="video/mp4" autoplay></video>
-            </div>
-            <div class="main">
-                <video id="{{track.getParticipantId}}-fullscreen" type="video/mp4" autoplay></video>
-            </div>
-        </div>
-    </div>
-</template>
-
 <template name="meetLowLevel">
     {{#if isActive}}
     <div class="remote-users">
