+++ core/modules/meet/client/meet-low-level.js
@@ -1,7 +1,133 @@
-let meetJs
+import { getTrackType, updateTrack, attachLocalTracks, replaceLocalTrack } from './tracks/utils'
+
+window.addEventListener('DOMContentLoaded', () => {
+    jitsiMeetJS = undefined
+    if (!Meteor.settings.public.meet) return
+
+    const head = document.querySelector('head')
+
+    const scriptLowLevel = document.createElement('script')
+    scriptLowLevel.src = `https://${Meteor.settings.public.meet.serverURL}/libs/lib-jitsi-meet.min.js`
+    head.appendChild(scriptLowLevel)
+
+    scriptLowLevel.onload = () => {
+        meetLowLevel = window.JitsiMeetJS
+    }
+})
+
+Template.meetLowLevel.onCreated(function () {
+    this.avatarURL = new ReactiveVar()
+    this.connection = new ReactiveVar(undefined)
+    this.localTracks = new ReactiveVar([])
+    this.remoteTracks = new ReactiveVar({})
+
+    this.connectionStarted = false
+    this.room = undefined
+    this.roomName = undefined
+    this.usersIdsInCall = []
+
+    this.autorun(() => {
+        if (!Meteor.userId()) return
+
+        const user = Meteor.user({ fields: { 'profile.avatar': 1 } })
+
+        if (user) this.avatarURL.set(generateRandomAvatarURLForUser(user))
+    })
+
+    this.autorun(() => {
+        const user = Meteor.user({ fields: { 'profile.videoRecorder': 1 } })
+
+        if (!user || !meetLowLevel) return
+        meetLowLevel
+            .createLocalTracks({
+                devices: ['video'],
+                cameraDeviceId: user?.profile?.videoRecorder,
+            })
+            .then((tracks) => replaceLocalTrack(this, tracks[0]))
+            .catch((err) => console.error('An error occured while creating local tracks', err))
+    })
+
+    this.autorun(() => {
+        const user = Meteor.user({ fields: { 'profile.audioRecorder': 1 } })
+
+        if (!user || !meetLowLevel) return
+        meetLowLevel
+            .createLocalTracks({
+                devices: ['audio'],
+                micDeviceId: user?.profile?.audioRecorder,
+            })
+            .then((tracks) => replaceLocalTrack(this, tracks[0]))
+            .catch((err) => console.error('An error occured while creating local tracks', err))
+    })
+
+    window.addEventListener(eventTypes.onUsersComeCloser, (e) => onUsersComeCloser(e, this))
+    window.addEventListener(eventTypes.onUsersMovedAway, (e) => onUsersMovedAway(e, this))
+    window.addEventListener(eventTypes.onUserPropertyUpdated, (e) => onUserPropertyUpdated(e, this))
+})
+
+Template.meetLowLevel.onDestroyed(() => {
+    window.removeEventListener(eventTypes.onUsersComeCloser, (e) => onUsersComeCloser(e, this))
+    window.removeEventListener(eventTypes.onUsersMovedAway, (e) => onUsersMovedAway(e, this))
+    window.removeEventListener(eventTypes.onUserPropertyUpdated, (e) => onUserPropertyUpdated(e, this))
+})
+
+Template.meetLowLevel.helpers({
+    avatarURL() {
+        return Template.instance().avatarURL.get()
+    },
+    isActive() {
+        return Template.instance().connection.get() !== undefined && Template.instance().usersIdsInCall.length > 0
+    },
+    remoteTracks() {
+        console.log(
+            '🚀 ---------------------------------------------------------------------------------------------------------------------------------🚀'
+        )
+        console.log(
+            '🚀 - file: meet-low-level.js:551 - remoteTracks - Template.instance().remoteTracks.get():',
+            Template.instance().remoteTracks.get()
+        )
+        console.log(
+            '🚀 ---------------------------------------------------------------------------------------------------------------------------------🚀'
+        )
+        // Check a better way to remove undefined tracks
+        return Object.values(Template.instance().remoteTracks.get()).filter((track) => track.audio && track.camera)
+    },
+    isLocalVideoActive() {
+        const user = Meteor.user({ fields: { 'profile.shareVideo': 1 } })
+
+        if (!user || !user.profile) return
+        return user.profile.shareVideo && Template.instance().localTracks.get().length > 0
+    },
+    isSharingScreen() {
+        const localTracks = Template.instance().localTracks.get()
+
+        return localTracks?.find((t) => t.getType() === 'video' && t.getVideoType() === 'desktop')
+    },
+    fullscreenTrack() {
+        const fullscreenTrackId = Session.get('fullscreenTrackId')
+
+        if (!Session.get('fullscreenTrackId')) return
+        let fullscreenTrack
+        const remoteTracks = Object.values(Template.instance().remoteTracks.get())
+
+        remoteTracks?.forEach((tracks) => {
+            const tracksMap = Object.values(tracks)
+            tracksMap.forEach((track) => {
+                if (track && typeof track !== 'string' && track.getId() === fullscreenTrackId) {
+                    fullscreenTrack = track
+                }
+            })
+        })
+
+        return fullscreenTrack
+    },
+})
+
+/*
+ ** LowMeetJs
+ */
 
 const DOMAIN = '8x8.vc'
-
 const getOptions = (roomName) => ({
     // Connection
     hosts: {
@@ -9,7 +135,6 @@ const getOptions = (roomName) => ({
         muc: `conference.${DOMAIN}`,
         focus: `focus.${DOMAIN}`,
     },
-    // serviceUrl: `https://${DOMAIN}/http-bind?room=${'laa'}`,
     serviceUrl: `wss://${DOMAIN}/xmpp-websocket?room=${roomName}`,
     websocketKeepAliveUrl: `https://${DOMAIN}/_unlock?room=${roomName}`,
 
@@ -50,29 +175,78 @@ const getOptions = (roomName) => ({
     __end: true,
 })
 
-window.addEventListener('DOMContentLoaded', () => {
-    if (!Meteor.settings.public.meet) return
+const connect = async (template) => {
+    console.log('Connection started')
 
-    const head = document.querySelector('head')
+    if (!template.connection.get()) {
+        const options = getOptions(template.roomName)
+        const user = Meteor.user({ fields: { 'profile.audioRecorder': 1, 'profile.videoRecorder': 1 } })
 
-    const scriptLowLevel = document.createElement('script')
-    scriptLowLevel.src = `https://${Meteor.settings.public.meet.serverURL}/libs/lib-jitsi-meet.min.js`
-    head.appendChild(scriptLowLevel)
+        meetLowLevel.init(options)
+        meetLowLevel.setLogLevel(meetLowLevel.logLevels.ERROR)
 
-    scriptLowLevel.onload = () => {
-        meetJs = window.JitsiMeetJS
+        await meetLowLevel
+            .createLocalTracks({
+                devices: ['audio', 'video'],
+                cameraDeviceId: user?.profile?.videoRecorder,
+                micDeviceId: user?.profile?.audioRecorder,
+            })
+            .then((tracks) => {
+                updateTrack('video', tracks)
+                updateTrack('audio', tracks)
+                attachLocalTracks(tracks)
+                template.localTracks.set(tracks)
+            })
+            .catch((err) => console.error('An error occured while creating local tracks', err))
+
+        const connection = new meetLowLevel.JitsiConnection(null, null, options)
+
+        connection.addEventListener(meetLowLevel.events.connection.CONNECTION_ESTABLISHED, () =>
+            onConnectionSuccess(template)
+        )
+        connection.addEventListener(meetLowLevel.events.connection.CONNECTION_FAILED, onConnectionFailed)
+        connection.addEventListener(meetLowLevel.events.connection.CONNECTION_DISCONNECTED, () =>
+            onConnectionDisconnected(template)
+        )
+
+        connection.connect()
+        template.connection.set(connection)
     }
-})
+}
 
-const getTrackId = (track) => `${track.getParticipantId()}-${getTrackType(track)}`
+const disconnect = async (template) => {
+    console.log('DISCONNECT')
 
-const getTrackType = (track) => {
-    if (track.getType() === 'audio') return 'audio'
-    else if (track.getType() === 'video') return track.getVideoType()
+    template.connectionStarted = false
+    Meteor.users.update(Meteor.userId(), {
+        $unset: { 'profile.meetRoomName': 1 },
+    })
+
+    if (template.room?.room) {
+        try {
+            template.room.leave()
+        } catch (err) {
+            console.log('Error during leaving', err)
+        }
+    }
+
+    const connection = template.connection.get()
+
+    connection.disconnect()
+    connection.removeEventListener(meetLowLevel.events.connection.CONNECTION_ESTABLISHED, (template) =>
+        onConnectionSuccess(template)
+    )
+    connection.removeEventListener(meetLowLevel.events.connection.CONNECTION_FAILED, onConnectionFailed)
+    connection.removeEventListener(meetLowLevel.events.connection.CONNECTION_DISCONNECTED, () =>
+        onConnectionDisconnected(template)
+    )
+
+    template.room = undefined
+    template.connection.set(undefined)
 }
 
 /*
- **  MeetJs events
+ **  LowMeetJs events listeners
  */
 const onTrackAdded = (template, track) => {
     // Since we attach local tracks separately, we do not need attach it again
@@ -148,14 +322,14 @@ const onConnectionSuccess = (template) => {
         }
 
         // Setup event listeners
-        template.room.on(meetJs.events.conference.TRACK_ADDED, (track) => onTrackAdded(template, track))
-        template.room.on(meetJs.events.conference.TRACK_REMOVED, (track) => onTrackRemoved(template, track))
-        template.room.on(meetJs.events.conference.CONFERENCE_JOINED, () => onConferenceJoined(template))
-        template.room.on(meetJs.events.conference.CONFERENCE_LEFT, onConferenceLeft)
-        template.room.on(meetJs.events.conference.USER_JOINED, (userId, participant) =>
+        template.room.on(meetLowLevel.events.conference.TRACK_ADDED, (track) => onTrackAdded(template, track))
+        template.room.on(meetLowLevel.events.conference.TRACK_REMOVED, (track) => onTrackRemoved(template, track))
+        template.room.on(meetLowLevel.events.conference.CONFERENCE_JOINED, () => onConferenceJoined(template))
+        template.room.on(meetLowLevel.events.conference.CONFERENCE_LEFT, onConferenceLeft)
+        template.room.on(meetLowLevel.events.conference.USER_JOINED, (userId, participant) =>
             onUserJoined(template, userId, participant)
         )
-        template.room.on(meetJs.events.conference.USER_LEFT, (id) => console.log('user left!', id))
+        template.room.on(meetLowLevel.events.conference.USER_LEFT, (id) => console.log('user left!', id))
 
         // Join
         template.room.setDisplayName(user?.profile?.name)
@@ -181,243 +355,12 @@ const onConnectionDisconnected = (template) => {
     template.localTracks.set([])
     template.remoteTracks.set({})
     template.usersInCall = []
-    toggleUserProperty('shareScreen')
+    toggleUserProperty('shareScreen', false)
 }
 
 /*
- ** Video/Audio Track templates
+ ** Events listeners
  */
-
-const updateTrackMuteState = (track) => {
-    if (template.isMuted) {
-        template.isMuted.set(track.isMuted())
-    }
-}
-
-const trackAttach = (template, trackId) => {
-    const { track } = Template.currentData() // Do not replace it to use 'template' props, need use global 'Template' to get the track
-
-    if (!track) return
-    const el = document.getElementById(trackId || getTrackId(track))
-
-    if (template.isMuted) {
-        template.isMuted.set(track.isMuted())
-    }
-
-    track.addEventListener(meetJs.events.track.TRACK_MUTE_CHANGED, () => updateTrackMuteState(track))
-
-    if (!el) return
-    track.attach(el)
-}
-
-const trackDetach = (trackId) => {
-    const { track } = Template.currentData()
-
-    if (!track) return
-    const el = document.getElementById(trackId || getTrackId(track))
-
-    track.removeEventListener(meetJs.events.track.TRACK_MUTE_CHANGED, () => updateTrackMuteState(track))
-
-    if (!el) return
-    track.detach(el)
-}
-
-Template.remoteVideoTrack.onCreated(function () {
-    this.isMuted = new ReactiveVar(false)
-})
-Template.remoteAudioTrack.onCreated(function () {
-    this.isMuted = new ReactiveVar(false)
-})
-Template.remoteDesktopTrack.onCreated(function () {
-    this.isMuted = new ReactiveVar(false)
-})
-Template.remoteFullscreen.onCreated(function () {
-    this.isMuted = new ReactiveVar(false)
-})
-
-Template.remoteVideoTrack.onRendered(function () {
-    this.autorun(() => trackAttach(this))
-})
-Template.remoteAudioTrack.onRendered(function () {
-    this.autorun(() => trackAttach(this))
-})
-Template.remoteDesktopTrack.onRendered(function () {
-    this.autorun(() => trackAttach(this))
-})
-Template.remoteFullscreen.onRendered(function () {
-    this.autorun(() => {
-        const { track } = Template.currentData()
-        trackAttach(this, `${track.getParticipantId()}-fullscreen`)
-        trackAttach(this, `${track.getParticipantId()}-fullscreenBackground`)
-    })
-})
-
-Template.remoteVideoTrack.onDestroyed(function () {
-    this.autorun(() => trackDetach())
-})
-Template.remoteAudioTrack.onDestroyed(function () {
-    this.autorun(() => trackDetach())
-})
-Template.remoteDesktopTrack.onDestroyed(function () {
-    this.autorun(() => trackDetach())
-})
-Template.remoteFullscreen.onDestroyed(function () {
-    this.autorun(() => {
-        const { track } = Template.currentData()
-        trackDetach(`${track.getParticipantId()}-fullscreen`)
-        trackDetach(`${track.getParticipantId()}-fullscreenBackground`)
-    })
-})
-
-Template.remoteVideoTrack.events({
-    'click .webcam': function (event) {
-        event.preventDefault()
-        event.stopPropagation()
-        const { track } = Template.currentData()
-
-        if (!track) return
-        Session.set('fullscreenTrackId', Session.get('fullscreenTrackId') === track.getId() ? null : track.getId())
-    },
-})
-Template.remoteDesktopTrack.events({
-    'click .js-screenshare': function (event) {
-        event.preventDefault()
-        event.stopPropagation()
-        const { track } = Template.currentData()
-
-        if (!track) return
-        Session.set('fullscreenTrackId', Session.get('fullscreenTrackId') === track.getId() ? null : track.getId())
-    },
-})
-
-Template.remoteVideoTrack.helpers({
-    isMuted: () => Template.instance().isMuted.get(),
-    avatar: () =>
-        generateRandomAvatarURLForUser({
-            _id: 'usr_a',
-            profile: { name: 'Guest', avatar: 'cat' },
-        }),
-})
-Template.remoteAudioTrack.helpers({
-    isMuted: () => Template.instance().isMuted.get(),
-})
-Template.remoteDesktopTrack.helpers({
-    isMuted: () => Template.instance().isMuted.get(),
-})
-
-/*
- ** MeetLowLevel templates
- */
-
-const updateTrack = (type, tracks) => {
-    if (!Meteor.userId()) return
-
-    const user = Meteor.user({ fields: { 'profile.shareAudio': 1, 'profile.shareVideo': 1 } })
-    const track = tracks.find((t) => t.getType() === type)
-
-    if (!track) return
-    if (type === 'audio') user?.profile?.shareAudio ? track.unmute() : track.mute()
-    else if (type === 'video') user?.profile?.shareVideo ? track.unmute() : track.mute()
-}
-
-const replaceLocalTrack = (template, newTrack) => {
-    const localTracks = template.localTracks.get()
-
-    if (localTracks) {
-        template.localTracks.set(
-            localTracks.map((track) => {
-                if (track.getType() === newTrack.getType()) {
-                    // Replace old tracks by the new one
-                    template.room.replaceTrack(track, newTrack)
-                    return newTrack
-                }
-                return track
-            })
-        )
-        attachLocalTracks([newTrack])
-    }
-}
-
-const attachLocalTracks = (tracks) => {
-    tracks.forEach((track) => {
-        if (track.getType() === 'video') {
-            let videoNode = document.querySelector('#video-stream-me')
-            track.attach(videoNode)
-        } else if (track.getType() === 'audio') {
-            let audioNode = document.querySelector('#audio-stream-me')
-            // tracks[i].attach(audioNode)
-        }
-    })
-}
-
-const connect = async (template) => {
-    console.log('Connection started')
-
-    if (!template.connection.get()) {
-        const options = getOptions(template.roomName)
-        const user = Meteor.user({ fields: { 'profile.audioRecorder': 1, 'profile.videoRecorder': 1 } })
-
-        meetJs.init(options)
-        meetJs.setLogLevel(meetJs.logLevels.ERROR)
-
-        await meetJs
-            .createLocalTracks({
-                devices: ['audio', 'video'],
-                cameraDeviceId: user?.profile?.videoRecorder,
-                micDeviceId: user?.profile?.audioRecorder,
-            })
-            .then((tracks) => {
-                updateTrack('video', tracks)
-                updateTrack('audio', tracks)
-                attachLocalTracks(tracks)
-                template.localTracks.set(tracks)
-            })
-            .catch((err) => console.error('An error occured while creating local tracks', err))
-
-        const connection = new meetJs.JitsiConnection(null, null, options)
-
-        connection.addEventListener(meetJs.events.connection.CONNECTION_ESTABLISHED, () =>
-            onConnectionSuccess(template)
-        )
-        connection.addEventListener(meetJs.events.connection.CONNECTION_FAILED, onConnectionFailed)
-        connection.addEventListener(meetJs.events.connection.CONNECTION_DISCONNECTED, () =>
-            onConnectionDisconnected(template)
-        )
-
-        connection.connect()
-        template.connection.set(connection)
-    }
-}
-
-const disconnect = async (template) => {
-    console.log('DISCONNECT')
-
-    template.connectionStarted = false
-    Meteor.users.update(Meteor.userId(), {
-        $unset: { 'profile.meetRoomName': 1 },
-    })
-
-    if (template.room?.room) {
-        await template.room
-            .leave()
-            .then(() => console.log('Room leaved'))
-            .catch((err) => console.log('Error during leaving', err))
-    }
-
-    const connection = template.connection.get()
-
-    connection.disconnect()
-    connection.removeEventListener(meetJs.events.connection.CONNECTION_ESTABLISHED, (template) =>
-        onConnectionSuccess(template)
-    )
-    connection.removeEventListener(meetJs.events.connection.CONNECTION_FAILED, onConnectionFailed)
-    connection.removeEventListener(meetJs.events.connection.CONNECTION_DISCONNECTED, () =>
-        onConnectionDisconnected(template)
-    )
-
-    template.room = undefined
-    template.connection.set(undefined)
-}
 
 const onUserPropertyUpdated = async (e, template) => {
     const { propertyName, propertyValue } = e.detail
@@ -431,7 +374,7 @@ const onUserPropertyUpdated = async (e, template) => {
         updateTrack('video', localTracks)
     } else if (propertyName === 'shareScreen') {
         if (propertyValue) {
-            await meetJs.createLocalTracks({ devices: ['desktop'] }).then((tracks) => {
+            await meetLowLevel.createLocalTracks({ devices: ['desktop'] }).then((tracks) => {
                 const currentDesktopTrack = localTracks.find((t) => t.getVideoType() === 'desktop')
                 const screenNode = document.querySelector('#video-screen-me')
                 const track = tracks[0] // Since we just ask for desktop, we will only have one item
@@ -477,7 +420,7 @@ const onUsersMovedAway = (e, template) => {
 const onUsersComeCloser = (e, template) => {
     const { users } = e.detail
 
-    if (meetJs && !template.connection.get() && !template.connectionStarted) {
+    if (meetLowLevel && !template.connection.get() && !template.connectionStarted) {
         template.connectionStarted = true
         let roomName = users[0]?.profile?.meetRoomName
 
@@ -507,102 +450,3 @@ const onUsersComeCloser = (e, template) => {
         if (!template.usersIdsInCall.includes(user._id)) template.usersIdsInCall.push(user._id)
     })
 }
-
-Template.meetLowLevel.onCreated(function () {
-    this.localTracks = new ReactiveVar([])
-    this.remoteTracks = new ReactiveVar({})
-    this.avatarURL = new ReactiveVar()
-
-    this.roomName = undefined
-    this.connection = new ReactiveVar(undefined)
-    this.connectionStarted = false
-    this.room = undefined
-
-    this.usersIdsInCall = []
-
-    this.autorun(() => {
-        if (!Meteor.userId()) return
-
-        const user = Meteor.user({ fields: { 'profile.avatar': 1 } })
-
-        if (user) this.avatarURL.set(generateRandomAvatarURLForUser(user))
-    })
-
-    this.autorun(() => {
-        const user = Meteor.user({ fields: { 'profile.videoRecorder': 1 } })
-
-        if (!user || !meetJs) return
-        meetJs
-            .createLocalTracks({
-                devices: ['video'],
-                cameraDeviceId: user?.profile?.videoRecorder,
-            })
-            .then((tracks) => replaceLocalTrack(this, tracks[0]))
-            .catch((err) => console.error('An error occured while creating local tracks', err))
-    })
-
-    this.autorun(() => {
-        const user = Meteor.user({ fields: { 'profile.audioRecorder': 1 } })
-
-        if (!user || !meetJs) return
-        meetJs
-            .createLocalTracks({
-                devices: ['audio'],
-                micDeviceId: user?.profile?.audioRecorder,
-            })
-            .then((tracks) => replaceLocalTrack(this, tracks[0]))
-            .catch((err) => console.error('An error occured while creating local tracks', err))
-    })
-
-    window.addEventListener(eventTypes.onUsersComeCloser, (e) => onUsersComeCloser(e, this))
-    window.addEventListener(eventTypes.onUsersMovedAway, (e) => onUsersMovedAway(e, this))
-    window.addEventListener(eventTypes.onUserPropertyUpdated, (e) => onUserPropertyUpdated(e, this))
-})
-
-Template.meetLowLevel.onDestroyed(() => {
-    window.removeEventListener(eventTypes.onUsersComeCloser, (e) => onUsersComeCloser(e, this))
-    window.removeEventListener(eventTypes.onUsersMovedAway, (e) => onUsersMovedAway(e, this))
-    window.removeEventListener(eventTypes.onUserPropertyUpdated, (e) => onUserPropertyUpdated(e, this))
-})
-
-Template.meetLowLevel.helpers({
-    avatarURL() {
-        return Template.instance().avatarURL.get()
-    },
-    isActive() {
-        return Template.instance().connection.get() !== undefined && Template.instance().usersIdsInCall.length > 0
-    },
-    remoteTracks() {
-        // Check a better way to remove undefined tracks
-        return Object.values(Template.instance().remoteTracks.get()).filter((track) => track.audio && track.camera)
-    },
-    isLocalVideoActive() {
-        const user = Meteor.user({ fields: { 'profile.shareVideo': 1 } })
-
-        if (!user || !user.profile) return
-        return user.profile.shareVideo && Template.instance().localTracks.get().length > 0
-    },
-    isSharingScreen() {
-        const localTracks = Template.instance().localTracks.get()
-
-        return localTracks?.find((t) => t.getType() === 'video' && t.getVideoType() === 'desktop')
-    },
-    fullscreenTrack() {
-        const fullscreenTrackId = Session.get('fullscreenTrackId')
-
-        if (!Session.get('fullscreenTrackId')) return
-        let fullscreenTrack
-        const remoteTracks = Object.values(Template.instance().remoteTracks.get())
-
-        remoteTracks?.forEach((tracks) => {
-            const tracksMap = Object.values(tracks)
-            tracksMap.forEach((track) => {
-                if (track && typeof track !== 'string' && track.getId() === fullscreenTrackId) {
-                    fullscreenTrack = track
-                }
-            })
-        })
-
-        return fullscreenTrack
-    },
-})
