+++ core/modules/meet/client/tracks/remote-audio-track.js
@@ -1,0 +1,17 @@
+import { trackAttach, trackDetach } from './utils'
+
+Template.remoteAudioTrack.onCreated(function () {
+    this.isMuted = new ReactiveVar(false)
+})
+
+Template.remoteAudioTrack.onRendered(function () {
+    this.autorun(() => trackAttach(this))
+})
+
+Template.remoteAudioTrack.onDestroyed(function () {
+    this.autorun(() => trackDetach(this))
+})
+
+Template.remoteAudioTrack.helpers({
+    isMuted: () => Template.instance().isMuted.get(),
+})
