+++ core/modules/meet/client/tracks/remote-desktop-track.hbs.html
@@ -1,0 +1,5 @@
+<template name="remoteDesktopTrack">
+    <div class="js-screenshare screenshare-preview {{#if isMuted}}hide{{/if}}">
+        <video id="{{track.getParticipantId}}-desktop" type="video/mp4" autoplay muted></video>
+    </div>
+</template>
