+++ core/modules/meet/client/tracks/remote-video-track.scss
@@ -1,0 +1,111 @@
+@import '../../../../client/variables';
+
+.remote-video {
+    z-index: 5;
+    position: relative;
+    width: 280px;
+    height: 180px;
+    background-color: rgba($main-color, 0.9);
+    border-radius: 15px;
+    background-size: cover;
+    overflow-x: hidden;
+    margin-bottom: 10px;
+
+    &.talking {
+        border: 2px solid var(--secondary-color);
+    }
+
+    &::before {
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        color: white;
+    }
+
+    &[data-state='calling']::before {
+        content: 'Calling…';
+    }
+
+    &[data-state='media-error']::before {
+        content: 'Media error 😢';
+    }
+
+    .stream-name {
+        position: absolute;
+        left: 5px;
+        top: 5px;
+        background-color: rgba($main-color, 0.75);
+        border-radius: 4px;
+        padding: 3px 10px;
+        z-index: 10;
+        color: #e5e4e7;
+    }
+
+    img,
+    video {
+        pointer-events: none;
+    }
+
+    video {
+        width: 100%;
+        height: 180px;
+        border-radius: 10px;
+        object-fit: cover;
+    }
+
+    .webcam {
+        cursor: pointer;
+        width: 100%;
+        height: 100%;
+
+        img,
+        video {
+            width: 280px;
+            height: 180px;
+            position: absolute;
+            background-color: #292929;
+            background-image: url('loader-white.gif');
+            background-size: inherit;
+            background-repeat: no-repeat;
+            background-position: center;
+        }
+
+        img {
+            z-index: 8;
+            cursor: default;
+        }
+    }
+
+    .screenshare-preview {
+        cursor: pointer;
+        width: 128px;
+        height: 72px;
+        z-index: 12;
+        position: absolute;
+        border-radius: 6px;
+        bottom: 2px;
+        right: 2px;
+        border: 2px solid var(--primary-color);
+        background-color: #292929;
+        box-sizing: content-box;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        &:hover {
+            border: 2px solid var(--secondary-color-darken-20);
+        }
+
+        video {
+            width: 128px;
+            height: 72px;
+            background-color: #292929;
+            background-image: url('loader-white.gif');
+            background-size: 50%;
+            background-repeat: no-repeat;
+            background-position: center;
+            border-radius: 2px;
+        }
+    }
+}
